requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
boto3==1.35.75
botocore==1.35.75
build==1.2.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.11
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.7
deepl==1.19.1
Deprecated==1.2.14
distro==1.9.0
durationpy==0.7
fastapi==0.115.0
feedparser==6.0.11
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
JPype1==1.5.0
jq==1.8.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwipiepy==0.19.0
kiwipiepy_model==0.19.0
konlpy==0.6.0
kubernetes==31.0.0
langchain==0.3.10
langchain-aws==0.2.9
langchain-community==0.3.10
langchain-core==0.3.32
langchain-openai==0.2.1
langchain-teddynote==0.3.42
langchain-text-splitters==0.3.0
langgraph==0.2.68
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.43
langsmith==0.1.129
lxml==5.3.0
lz4==4.3.3
markdown-it-py==3.0.0
marshmallow==3.22.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.19.2
openai==1.50.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
pandas==2.2.3
pdf2image==1.17.0
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.9.0
posthog==3.6.6
protobuf==4.25.5
protoc-gen-openapiv2==0.0.1

pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4

PyPika==0.48.9
pyproject_hooks==1.2.0

python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2

rank-bm25==0.2.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
s3transfer==0.10.4
sgmllib3k==1.0.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
sympy==1.13.3
tavily-python==0.5.0
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.20.0
tqdm==4.66.5
typer==0.12.5
types-requests==2.32.0.20240914
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
wget==3.2
wrapt==1.16.0
yarl==1.13.1
ipython==8.12.0
